requirements.txt
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
Pillow>=9.0.0
python-decouple>=3.6
drf-spectacular>=0.27.0
psycopg2-binary>=2.9.0
dj-database-url>=2.0.0
gunicorn>=21.0.0
django-ckeditor>=6.7.0
